requirements.txt
# 可以通过运行【install_dependencies.py】文件来安装依赖
# cd 目录：如下
# cd e:\git_project\comfyui-xnantool
# python install_dependencies.py
# 检测
# cd e:\git_project\comfyui-xnantool
# python verify_installation_xnantool.py

# 基础通用依赖

torch
torchvision  # 添加torchvision，YOLO需要它进行图像处理
numpy

# YOLO相关依赖
ultralytics>=8.0.0  # 指定版本要求，确保兼容性
matplotlib>=3.5.0
requests>=2.28.0
pillow>=9.0.0  # pillow是PIL的更新版本，无需重复指定PIL
opencv-python>=4.5.0  # 添加OpenCV，YOLO检测常用的图像处理库

# SAM相关依赖
# git+https://github.com/facebookresearch/segment-anything.git  # SAM模型库
# git+https://gitcode.com/GitHub_Trending/se/segment-anything.git # SAM模型库加速镜像
# 可以通过秋叶命令提示符直接python install git+https://gitcode.com/GitHub_Trending/se/segment-anything.git
onnxruntime  # ONNX运行时，用于模型导出
onnx  # ONNX格式支持
# pycocotools-windows-fix; sys_platform == 'win32'  # Windows系统的pycocotools替代方案
# pycocotools; sys_platform != 'win32'  # 非Windows系统的pycocotools

# 魔搭api相关依赖
openai
# requests
#PIL # pillow是PIL的更新版本，无需重复指定PIL
# httpx
# socksio

# Ollama相关依赖
ollama

# 视频处理相关依赖
imageio>=2.9.0  # 用于读取和写入图像序列，包括GIF
imageio-ffmpeg>=0.4.2  # imageio的ffmpeg插件，用于视频处理
soundfile>=0.12.0  # 用于处理音频文件，支持音频合并功能
# 注意：本插件的视频转音频功能需要系统安装ffmpeg命令行工具
# Windows用户可以从 https://ffmpeg.org/download.html 下载并安装
# 确保ffmpeg命令在系统PATH环境变量中可用
# 如果是秋叶包，记得卸载试下运行，不行然后再安装后点同步ffmpeg

# Markdown转Excel功能依赖
pandas>=1.5.0
openpyxl>=3.0.0
markdown>=3.3.0

# 加密解密相关
cryptography>=3.4.8
pycryptodome>=3.18.0

# 火山引擎SDK（用于豆包Seedream图像生成和豆包Seedance视频生成）
volcengine-python-sdk[ark]>=1.0.0
